+++ src/main/java/appland/settings/AppMapProjectSettings.java
@@ -1,5 +1,6 @@
 package appland.settings;
 
+import com.google.common.collect.Lists;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -15,6 +16,25 @@ public class AppMapProjectSettings {
     private String activeRecordingURL;
     @Nullable
     private String recentAppMapStorageLocation;
+    @Nullable
+    private String cloudServerUrl;
+    private boolean confirmAppMapUpload = true;
+
+    public AppMapProjectSettings() {
+    }
+
+    /**
+     * Copy constructor.
+     *
+     * @param settings The settings to copy.
+     */
+    public AppMapProjectSettings(@NotNull AppMapProjectSettings settings) {
+        this.recentRemoteRecordingURLs = settings.recentRemoteRecordingURLs == null ? null : Lists.newLinkedList(settings.recentRemoteRecordingURLs);
+        this.activeRecordingURL = settings.activeRecordingURL;
+        this.recentAppMapStorageLocation = settings.recentAppMapStorageLocation;
+        this.cloudServerUrl = settings.cloudServerUrl;
+        this.confirmAppMapUpload = settings.confirmAppMapUpload;
+    }
 
     @NotNull
     public synchronized String getRecentAppMapStorageLocation() {
@@ -56,13 +76,21 @@ public class AppMapProjectSettings {
         this.activeRecordingURL = activeRecordingURL;
     }
 
-    @Override
-    public String toString() {
-        return "AppMapProjectSettings{" +
-                "recentRemoteRecordingURLs=" + recentRemoteRecordingURLs +
-                ", activeRecordingURL='" + activeRecordingURL + '\'' +
-                ", recentAppMapStorageLocation='" + recentAppMapStorageLocation + '\'' +
-                '}';
+    public synchronized boolean getConfirmAppMapUpload() {
+        return confirmAppMapUpload;
+    }
+
+    public synchronized void setConfirmAppMapUpload(boolean confirmAppMapUpload) {
+        this.confirmAppMapUpload = confirmAppMapUpload;
+    }
+
+    @Nullable
+    public synchronized String getCloudServerUrl() {
+        return cloudServerUrl;
+    }
+
+    public synchronized void setCloudServerUrl(@Nullable String cloudServerUrl) {
+        this.cloudServerUrl = cloudServerUrl;
     }
 
     @Override
@@ -70,11 +98,22 @@ public class AppMapProjectSettings {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         AppMapProjectSettings that = (AppMapProjectSettings) o;
-        return Objects.equals(recentRemoteRecordingURLs, that.recentRemoteRecordingURLs) && Objects.equals(activeRecordingURL, that.activeRecordingURL) && Objects.equals(recentAppMapStorageLocation, that.recentAppMapStorageLocation);
+        return confirmAppMapUpload == that.confirmAppMapUpload && Objects.equals(recentRemoteRecordingURLs, that.recentRemoteRecordingURLs) && Objects.equals(activeRecordingURL, that.activeRecordingURL) && Objects.equals(recentAppMapStorageLocation, that.recentAppMapStorageLocation) && Objects.equals(cloudServerUrl, that.cloudServerUrl);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(recentRemoteRecordingURLs, activeRecordingURL, recentAppMapStorageLocation);
+        return Objects.hash(recentRemoteRecordingURLs, activeRecordingURL, recentAppMapStorageLocation, cloudServerUrl, confirmAppMapUpload);
+    }
+
+    @Override
+    public String toString() {
+        return "AppMapProjectSettings{" +
+                "recentRemoteRecordingURLs=" + recentRemoteRecordingURLs +
+                ", activeRecordingURL='" + activeRecordingURL + '\'' +
+                ", recentAppMapStorageLocation='" + recentAppMapStorageLocation + '\'' +
+                ", cloudServerUrl='" + cloudServerUrl + '\'' +
+                ", confirmAppMapUpload=" + confirmAppMapUpload +
+                '}';
     }
 }