+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # AppMap Plugin Changelog
 
+## [0.6.0]
+### Added
+- Uploading AppMaps to AppMap Cloud
+
+### Changed
+- Updated AppMap JavaScript components
+
 ## [0.5.1]
 ### Added
-- Warning displayed when JCEF not available in JDK
+- Display a notification for AppMap JSON editors if the HTML application can't be displayed.
+- New feature to upload AppMaps to a remote server.
 
 ### Changed
 - Improved search in the AppMaps tool window